20190421
Continue update animation image view
