+++ Example/DarkEggKit/AnimationImageViewController.swift
@@ -11,21 +11,24 @@ import UIKit
 import DarkEggKit
 
 class AnimationImageViewController: UIViewController {
-    @IBOutlet weak var aniImgView: AImageView!
+    @IBOutlet weak var aniImgView_1: AnimationImageView!
+    @IBOutlet weak var aniImgView_2: AnimationImageView!
+    @IBOutlet weak var aniImgView_3: AnimationImageView!
+    @IBOutlet weak var aniImgView_4: AnimationImageView!
 
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
-        aniImgView.backgroundColor = .white;
+        //aniImgView.backgroundColor = .white;
         
         //let data = try! Data(contentsOf: URL(fileReferenceLiteralResourceName: "colon_animation.png"))
         //let apng = AImage(data: data)
         
         Logger.debug("Start")
         var apng: AnimationImage? // = AImage(url: "https://apng.onevcat.com/assets/elephant.png")
-        switch Int.random(in: 1...4) {
+        switch Int.random(in: 4...4) {
         case 1:
-            apng = AnimationImage(url: "http://ics-web.jp/lab-data/140930_apng/images/elephant_apng_zopfli.png")
+            apng = AnimationImage(url: "https://flif.info/example-animation/spinfox_50.png")
         case 2:
             apng = AnimationImage(url: "https://orig00.deviantart.net/df6e/f/2012/287/f/8/i_want_to_be_a_hero__apng_animated__by_tamalesyatole-d5ht8eu.png")
         case 3:
@@ -40,10 +43,34 @@ class AnimationImageViewController: UIViewController {
         //let apng = AImage(url: "https://pbs.twimg.com/media/D2pMC1nVAAA2VQn.jpg")
         //let apng = AImage(url: "https://media.tenor.com/images/39fe167bdab90223bcc890bcb067b761/tenor.gif")
         //https://media2.giphy.com/media/xUPGcLbQtBlro61bdm/giphy.gif
-        aniImgView.aImage = apng
+        
+        let dummyImg = UIImage(named: "pipboy_01")
+        
+        aniImgView_1.aImage = apng
         //aniImgView.placeHolder = UIImage(named: "")
-        aniImgView.repeatMode = .infinite
-        aniImgView.delegate = self
+        aniImgView_1.repeatMode = .infinite
+        aniImgView_1.delegate = self
+        //
+        aniImgView_2.aImage = apng
+        //aniImgView_2.image = dummyImg
+        //aniImgView_2.placeHolder = dummyImg
+        aniImgView_2.repeatMode = .infinite
+        aniImgView_2.delegate = self
+        //
+        aniImgView_3.aImage = apng
+        //aniImgView_3.image = dummyImg
+        //aniImgView_3.placeHolder = UIImage(named: "")
+        aniImgView_3.repeatMode = .infinite
+        aniImgView_3.willShowProgress = false
+        aniImgView_3.delegate = self
+        //
+        aniImgView_4.aImage = apng
+        //aniImgView_4.image = dummyImg
+        //aniImgView_4.placeHolder = UIImage(named: "")
+        aniImgView_4.repeatMode = .infinite
+        aniImgView_4.delegate = self
+        //aniImgView_4.cancelDownloading()
+        
         Logger.debug("End")
     }
     
@@ -52,12 +79,18 @@ class AnimationImageViewController: UIViewController {
     }
 }
 
-extension AnimationImageViewController: AImageViewDelegate {
+extension AnimationImageViewController: AnimationImageViewDelegate {
 //    func animatedImageView(_ imageView: AImageView, didPlayAnimationLoops count: UInt) {
 //        Logger.debug()
 //    }
     
-    func animatedImageViewDidFinishAnimating(_ imageView: AImageView) {
+    func animatedImageViewDidFinishAnimating(_ imageView: AnimationImageView) {
         Logger.debug()
+    }
+}
+
+extension AnimationImageViewController {
+    @IBAction func onClearCacheButtonClicked(_ sender: UIButton) {
+        AnimationImageView.clearCache()
     }
 }
