+++ Example/DarkEggKit/PopupMessageViewController.swift
@@ -24,6 +24,15 @@ class PopupMessageViewController: UIViewController {
         //let image = UIImage(named: "dz_icon_info", in: Bundle(for: DZPopupMessageView.self), compatibleWith: nil)?.withRenderingMode(.alwaysTemplate)
         //self.imageView.image = image
     }
+    
+    override func viewDidDisappear(_ animated: Bool) {
+        Logger.debug()
+        self.imageView.image = UIImage()
+    }
+    
+    deinit {
+        Logger.debug()
+    }
 }
 
 extension PopupMessageViewController {
